% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snif.R
\name{print.snif}
\alias{print.snif}
\title{Print \code{snif} objects}
\usage{
\method{print}{snif}(x, ...)
}
\arguments{
\item{x}{An object of type "snif"}

\item{...}{Additional parameters to pass onto \code{print}}
}
\description{
\code{print} method for class "snip"
}
