% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{snif.df}
\alias{snif.df}
\title{Synthetic data for snif}
\format{A data.frame with 500 observations on 11 variables:}
\usage{
snif.df
}
\description{
Synthetic data for snif
}
\keyword{datasets}
