% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snif.R
\name{snif}
\alias{snif}
\title{Selection of nonlinear interactions by a forward stepwise algorithm}
\usage{
snif(formula, df, score = "BIC", degree = 3, maxit = ncol(df),
  main.only = NULL, linear.only = NULL)
}
\arguments{
\item{formula}{An initial formula for \code{snif} to start with.
\code{formula} supports linear variables, basis spline expansions via
\code{bs}, and interaction terms via \code{:}. It is highly recomended to
read the documentation before trying anything fancy with \code{formula}}

\item{df}{A data.frame containing the data}

\item{score}{A character argument that specifies the kind of scoring method
used to determine which variable to add to the model. supported options
are "BIC" (default), "AIC", and "PV" (p.value)}

\item{degree}{Degree of basis spline expansion for nonlinear effects}

\item{maxit}{Number of iterations to run. Default is \code{ncol(df)}}

\item{main.only}{Character vector of variables that are only considered for
main effects}

\item{linear.only}{Character vector of variables that are only considered to
have only linear effects}
}
\description{
\code{snif} is a greedy forward stepwise selection algorithm that takes
nonlinear effects and interactions into account when adding variables.
\code{snif} follows the strong hereditary principle (for X1:X2 to be
included as an interaction, both X1 and X2 both must have had main effects
selected).
}
\references{
Narisetty, Naveen N. and Mukherjee, Bhramar and Chen, Yin-Hsiu and Gonzalez,
 Richard and Meeker, John D. Selection of nonlinear interactions by a
 forward stepwise algorithm: Application to identifying environmental
 chemical mixtures affecting health outcomes. Statistics in Medicine.
 2019;38(9):1582-1600. \href{https://doi.org/10.1002/sim.8059}{10.1002/sim.8059}
}
\author{
Alexander Rix
}
