% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snif.R
\name{summary.snif}
\alias{summary.snif}
\title{Summarizing SNIF Fits}
\usage{
\method{summary}{snif}(object, alpha = 0.05, ...)
}
\arguments{
\item{object}{An object of type "snif"}

\item{alpha}{If method == "PV", \code{alpha} determines the selection criteon
for picking the best model.}

\item{...}{Additional arguments to \code{summary.lm} or \code{summary.glm},
depending on the type of model fit}
}
\description{
\code{summary} method for class "snif"
}
\details{
\code{summary.snif} extracts the best scoring model from \code{object} and
returns the linear model summary.
}
