% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snif.R
\name{predict.snif}
\alias{predict.snif}
\title{Predict Using The Best SNIF Fit}
\usage{
\method{predict}{snif}(object, newdata, ...)
}
\arguments{
\item{object}{An object of type "snif"}

\item{newdata}{An optional data frame in which to look for variables with
which to predict.  If omitted, the fitted values are used.}

\item{...}{Additional arguments to \code{predict.lm}}
}
\description{
predict method for class ‘"snif"’
}
\details{
\code{predict.snif} extracts the best scoring model from \code{object} and
returns the predictions from \code{predict.lm}.
}
