% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snif.R
\name{summary.snif}
\alias{summary.snif}
\title{Summarizing SNIF Fits}
\usage{
\method{summary}{snif}(object, ...)
}
\arguments{
\item{object}{An object of type "snif"}

\item{...}{Additional arguments to \code{summary.lm}}
}
\description{
‘summary’ method for class ‘"snif"’
}
\details{
\code{summary.snif} extracts the best scoring model from \code{object} and
returns the linear model summary.
}
